Breaking changes: Change the default value for pager.next_body_transformer = null / Refactor option usages
